poc/poc-11-pdf-parsing/requirements.txt
# POC-11 dependencies
pymupdf>=1.24.0
pdfplumber>=0.11.0
pdfminer.six>=20231228
pytesseract>=0.3.10
Pillow>=10.0
pydantic>=2.0
fastapi>=0.111
uvicorn>=0.30
httpx>=0.27
numpy>=1.26
python-multipart>=0.0.9
